Added Collapsable Leaflet Control mixin. Used by the History Box now. Extracted some X and Shrink button/links. Could create a more general class and more subclasses later. Made it operate a bit more nicely with the init hook. May mature later to include into other widgets more easily. Added History Box clear link.
With the recent addition of the collapsing History Box, re-configured the minimizing control system.
Re-ordered classes to get the context menu's events to trigger first / take priority, and cancel the others if it is open.
Thought I needed to extract the build context menu function.  Figured out the important part was when the objects were being initialized with the rest of the zMap UI, but I left it out to be more organized anyway.  I'm sure it can be cleaned up more later too in several ways.
Kept all listeners at the root document level in case some were not focused, which kind of seems impossible in some cases or at all for some controls, and is generally convenient.
Found a new `stopImmediatePropagation` command I have been using to good effect now for all of that.

It may be better to distribute the handlers more diversely over more specific DOM elements, but I would not be able to achieve all of the effects/situations mentioned above, AND there was a complex or impossible case that I could not operate in correctly.  I think it was that the main control would not collapse, in essence, toggling twice immediately.  If that could be fixed in a less desirable way, then other cases would be ambiguous and hard or impossible to deal with, such as having the context menu retain exclusive priority. whew.

Also changed the context menu to use the jQuery event binding framework like the other classes because it was being 'stomped out' when mixed, no longer having priority.

Re-configured the InfoBox title margins so it only affects the title.  This gave more room available for the expand button also in the header now.

Minor ZControl fix.  Nothing may use this but future-proof and small so went ahead.

Fixed Jason's changes to disable the shrink button.
Now it is a config setting in case one would want it, and probably utilizes that instead of hard-commenting out code.

Finally set-up collapsing / minimizing controls based on config now too!

Had to overhaul HistroyBox a bit as the first control in the new system to have this minimize behavior.
Have more DOM code for the other side of alternative/minimized state of the object, with the link to maximize and go back to the normal/expanded side of the control.
Got Leaflet control mixins to work well with the init hook now too.

Collapsable gives some basic functions, without doing much extra like creating and/or adding the shrink button itself, something the implementing controls might want to do themselves in a custom manner.

In the meanwhile, also fixed the CSS from the list view refactor.  Removed the extra padding so the main control search entries look more normal again.

A little doc filling out.
Refactored the search field to use the new clear/X link/button object as well.
Fixed a not clearing bug.

Was able to get the expand button in the History Box to act like the previous one without it being weirdly shallow, raised to the top, no border, small click area, not floated to the right, etc., AND fixed the original one that ran out of width room, also fixing the original style to be more exact in the process.